Support for suport of older androids
First a webpack issue - // https://github.com/zloirock/core-js/issues/514 workaround not loading on android 4.4. As with much webpack stuff, the magic going on is not perfectly clear. This workaround works, although it's not clear from the comments that it is supposed to be neeed.

Second, it turns out that the fix to the ios notch issue where framework 7 (our older version) wasn't respoding correcty broke old androids where vh coords we set on body are not supported. We need to do that only on ios. To support that we add a bootstrap.js file to be run before other js files. Currently it puts a io-user-agent class on on the html element. integrating apps need to include lib/bootstrap.js